+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to Fig will be documented in this file.
 
-## Unreleased
+## [0.4.0] - 2017-05-23
 ### Added
 - Support for nested variable definitions in multi-level `include`s